Refactor authentication logic into dedicated AuthService
- Create new AuthService class to centralize authentication logic
- Move legacy API key and GitHub authentication flows into service methods
- Simplify extension activation authentication process
- Improve code organization and separation of concerns
- Maintain existing authentication and migration behaviors
