+++ client/src/extension.ts
@@ -16,11 +16,10 @@ import {
 import { registerCommands } from './common/commands';
 import { registerActions } from './common/actions';
 import { setupFileWatchers } from './common/utils/fileWatchers';
-import { authenticateUserWithGitHub, validateApiKey } from './common/auth/api';
 import { initializeTelemetry } from '../../shared/telemetry';
-import { SESSION_TOKEN_KEY, SESSION_USER, TELEMETRY_EVENTS, TELEMETRY_NOTIFICATION } from '../../shared/constants';
-import { UserSession } from './common/auth/github';
-import { AUTH_MESSAGES } from './common/constants/messages';
+import { SESSION_TOKEN_KEY, TELEMETRY_EVENTS, TELEMETRY_NOTIFICATION } from '../../shared/constants';
+import { AuthService } from './common/auth/authService';
+
 
 
 let client: LanguageClient;
@@ -34,77 +33,9 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	reporter.sendTelemetryEvent(TELEMETRY_EVENTS.EXTENSION_ACTIVATED);
 
-	const session: UserSession | undefined = context.globalState.get(SESSION_USER);
-	const legacyApiKey: string | undefined = context.globalState.get(COMMANDS.USER_API_KEY);
-	const today = new Date();
-	
-	if (legacyApiKey && !session) {
-		const apiKeySession = await validateApiKey(legacyApiKey);
-		if (!apiKeySession) {
-			vscode.window.showErrorMessage(AUTH_MESSAGES.INVALID_API_KEY);
-			return;
-		}
-
-		const expiresAt = new Date(apiKeySession.session.data?.expires_at || '');
-		if (!apiKeySession.session.data?.expires_at || isNaN(expiresAt.getTime())) {
-			console.error('Invalid expiration date from server');
-			vscode.window.showErrorMessage(AUTH_MESSAGES.INVALID_API_KEY);
-			return;
-		}
-
-		if (apiKeySession.migration_notice) {
-			console.log('Migration notice:', apiKeySession.migration_notice);
-		}
-
-		await context.globalState.update(SESSION_TOKEN_KEY, apiKeySession.token);
-		await context.globalState.update(SESSION_USER, apiKeySession);
-		
-		// Show migration message with days remaining
-		const daysLeft = Math.ceil((expiresAt.getTime() - today.getTime()) / (1000 * 60 * 60 * 24));
-		const migrateAction = "Sign in with GitHub";
-		const remindLater = "Remind me later";
-		
-		const response = await vscode.window.showInformationMessage(
-			AUTH_MESSAGES.LEGACY_API_KEY_MIGRATION + ` You have ${daysLeft} days to migrate.`,
-			migrateAction,
-			remindLater
-		);
-
-		if (response === migrateAction) {
-			const isAuthenticated = await authenticateUserWithGitHub(context);
-			if (isAuthenticated) {
-				await context.globalState.update(COMMANDS.USER_API_KEY, undefined);
-				reporter.sendTelemetryEvent(TELEMETRY_EVENTS.LEGACY_USER_MIGRATED);
-			}
-		} else {
-			reporter.sendTelemetryEvent(TELEMETRY_EVENTS.LEGACY_USER_POSTPONED, {
-				days_left: daysLeft.toString(),
-				user_api_key: legacyApiKey,
-			});
-		}
-	} else if (!session || !session.user.has_agreed_to_terms) {
-		const isAuthenticated = await authenticateUserWithGitHub(context);
-		if (!isAuthenticated) {
-			vscode.window.showWarningMessage(AUTH_MESSAGES.AUTHENTICATION_REQUIRED);
-			reporter.sendTelemetryErrorEvent(TELEMETRY_EVENTS.AUTHENTICATION_FAILED, {
-				reason: 'User did not authenticate',
-				user_id: session?.user.id || 'unknown',
-			});
-			console.log('Failed to authenticate user. Exiting extension.');
-			return;
-
-		}
-		
-		const updatedSession = context.globalState.get<UserSession>(SESSION_USER);
-		if (!updatedSession?.user.has_agreed_to_terms) {
-			vscode.window.showErrorMessage(AUTH_MESSAGES.MUST_AGREE_TO_TERMS);
-			reporter.sendTelemetryErrorEvent(TELEMETRY_EVENTS.TERMS_NOT_ACCEPTED, {
-				reason: 'User did not agree to terms',
-				user_id: updatedSession?.user.id || 'unknown',
-			});
-			console.log('User has not agreed to terms. Exiting extension.');
-			return;
-		}
+	const authService = new AuthService(context);
+	if (!await authService.validateAuth()) {
+		return;
 	}
 
 	const serverModule = context.asAbsolutePath(
